requirements.txt
Flask
Flask-WTF
WTForms
Flask-Login
Flask-SQLAlchemy
psycopg2-binary
Flask-Bcrypt
Flask-Mail
python-dotenv
Pillow
itsdangerous
email-validator
gunicorn